I had some time to fortify my mind during the night. Yesterday, KFAC making things worse took me by surprise. I need to look into it. The reason for the must be the hyperparameters. Probably the LR for the covariance matrix updates is too high. That is the first thing I need to fiddle with.
This whole thing is just too strange.

Maybe instead of doubting the bad results of KFAC, I should doubt the good results of the MI plastic RNN. What I am going to do is take out that regular plastic RNN and study it more in depth on a smaller pattern.

I had some time to debug my mind as well. There is no way that the gradients for the weight update modulation would be small and that KFAC would push it up. Rather the case would be that they could be too big in their original form.

KFAC should work - that is one thing I am 100% sure of.

To start things off, let me lower the learning rates for the covariance matrix. I will start there. I already tried that, but not in a principled manner.