11:30am.
```
The learning rate is 2 ** -15
At iteration 9 the cost and the sign accuracy are 22.86703 and 0.95/1.
At iteration 19 the cost and the sign accuracy are 22.29082 and 0.966/1.
At iteration 29 the cost and the sign accuracy are 20.95501 and 0.98/1.
At iteration 39 the cost and the sign accuracy are 20.21015 and 0.98/1.
At iteration 49 the cost and the sign accuracy are 19.26765 and 0.944/1.
At iteration 59 the cost and the sign accuracy are 19.84659 and 0.978/1.
At iteration 69 the cost and the sign accuracy are 18.65338 and 0.946/1.
At iteration 79 the cost and the sign accuracy are 15.35054 and 0.99/1.
At iteration 89 the cost and the sign accuracy are 14.54056 and 0.978/1.
At iteration 99 the cost and the sign accuracy are 14.58197 and 0.98/1.
At iteration 109 the cost and the sign accuracy are 14.53506 and 0.966/1.
At iteration 119 the cost and the sign accuracy are 11.45427 and 0.996/1.
At iteration 129 the cost and the sign accuracy are 12.19544 and 0.958/1.
At iteration 139 the cost and the sign accuracy are 15.48549 and 0.938/1.
At iteration 149 the cost and the sign accuracy are 13.61434 and 0.956/1.
At iteration 159 the cost and the sign accuracy are 12.03898 and 0.928/1.
At iteration 169 the cost and the sign accuracy are 11.81945 and 0.952/1.
At iteration 179 the cost and the sign accuracy are 13.5684 and 0.934/1.
At iteration 189 the cost and the sign accuracy are 13.32757 and 0.932/1.
At iteration 199 the cost and the sign accuracy are 8.53623 and 0.972/1.
```

I managed to get it up to this. In fact I can init it so it starts at 95%.

11:45am.

```
inl bias = {si=const one; i=const half; s=const half; c=const zero}
```

Let me go back to the original plastic RNN.

I'll set it to this.

```
inl bias = {si=const zero; i=const zero; s=const zero; c=const 0.01f32}
```

Does it start at 95% with this?