Let me stop here. I should have had breakfast over an hour ago.
My plan is in complete shambles now. I had a lot of ideas during the night, but these recent experiments that I did reveal everything.

As long as you allow recurrent Hebbian connections and add the inputs to them and set the `theta` to something like 0.0001, you will get nearly perfect sign accuracy on its own. KFAC does not work and SGD barely does either. `theta` cannot be learned without blowing up.

The only thing SGD is doing is pushing up the plasticity coefficients. When I discovered the MI architecture I got the sense that I had made a huge breakthrough, but now that I understand where the performance gain is coming from I am notably less happy.

KFAC which tries to do more than merely pushing up the `alpha` biases fails and that is pretty damning for this architecture.

There might be a point to Hebbian learning, but not like this.