"1:25pm. Done with breakfast.
1:45pm. I am just sitting aimlessly in this seat. I am so demolished.

1:55pm. I had a lot of trust in differentiable plasticity and now it all fell apart.

2:10pm. No, it was inevitable. I was the one who was wrong.

There is a point to Hebbian learning, but not like this. In order to grasp something, absolute control is necessary.

In the first place, there is no way to stabilize modulation for the weight updates when the constraints are so tight like this...

No rather, I am going at it the wrong way.

It is not like I got only failure. I did find something interesting. Namely it seems that it blows up...

Hmmm...let me try something...

```
H = inl {H theta out input} -> H + tanh theta * input * out - abs theta * out * out * H
```

Let me multiply theta here by 0.0001f32. I'll consider this a part of initialization. If KFAC or something else corrects it then so be it.

2:25pm.

```
The learning rate is 2 ** -12
At iteration 9 the cost and the sign accuracy are 20.82334 and 0.99/1.
At iteration 19 the cost and the sign accuracy are 20.91031 and 0.992/1.
At iteration 29 the cost and the sign accuracy are 21.11125 and 0.98/1.
At iteration 39 the cost and the sign accuracy are 20.84579 and 0.966/1.
At iteration 49 the cost and the sign accuracy are 21.24944 and 0.922/1.
At iteration 59 the cost and the sign accuracy are 21.68553 and 0.918/1.
At iteration 69 the cost and the sign accuracy are 21.64828 and 0.904/1.
At iteration 79 the cost and the sign accuracy are 21.85255 and 0.906/1.
At iteration 89 the cost and the sign accuracy are 21.68585 and 0.904/1.
At iteration 99 the cost and the sign accuracy are 22.27941 and 0.9/1.
At iteration 109 the cost and the sign accuracy are 21.47608 and 0.9/1.
At iteration 119 the cost and the sign accuracy are 22.18732 and 0.884/1.
At iteration 129 the cost and the sign accuracy are 22.34362 and 0.878/1.
At iteration 139 the cost and the sign accuracy are 21.83654 and 0.886/1.
At iteration 149 the cost and the sign accuracy are 22.35347 and 0.86/1.
At iteration 159 the cost and the sign accuracy are 22.39102 and 0.866/1.
At iteration 169 the cost and the sign accuracy are 22.93355 and 0.844/1.
At iteration 179 the cost and the sign accuracy are 22.43116 and 0.876/1.
At iteration 189 the cost and the sign accuracy are 23.23186 and 0.842/1.
At iteration 199 the cost and the sign accuracy are 23.05609 and 0.852/1.
```

Once I add KFAC it immediately starts to worsen.

But you know, since I understand the secret of this now, maybe I could achieve something by focusing solely on modulation.

2:50pm.

```
The learning rate is 2 ** -10
At iteration 9 the cost and the sign accuracy are 20.85584 and 0.99/1.
At iteration 19 the cost and the sign accuracy are 20.88275 and 0.998/1.
At iteration 29 the cost and the sign accuracy are 20.97005 and 0.996/1.
At iteration 39 the cost and the sign accuracy are 20.86461 and 1/1.
At iteration 49 the cost and the sign accuracy are 20.90725 and 0.984/1.
At iteration 59 the cost and the sign accuracy are 21.03608 and 1/1.
At iteration 69 the cost and the sign accuracy are 20.80846 and 0.994/1.
At iteration 79 the cost and the sign accuracy are 20.85337 and 0.998/1.
At iteration 89 the cost and the sign accuracy are 20.6763 and 1/1.
At iteration 99 the cost and the sign accuracy are 20.82223 and 0.998/1.
At iteration 109 the cost and the sign accuracy are 20.77411 and 1/1.
At iteration 119 the cost and the sign accuracy are 20.61869 and 1/1.
At iteration 129 the cost and the sign accuracy are 20.73735 and 1/1.
At iteration 139 the cost and the sign accuracy are 20.85756 and 1/1.
At iteration 149 the cost and the sign accuracy are 20.89059 and 1/1.
At iteration 159 the cost and the sign accuracy are 21.11161 and 0.994/1.
At iteration 169 the cost and the sign accuracy are 20.74962 and 0.994/1.
At iteration 179 the cost and the sign accuracy are 20.66193 and 0.998/1.
At iteration 189 the cost and the sign accuracy are 21.13816 and 0.992/1.
At iteration 199 the cost and the sign accuracy are 20.92275 and 1/1.
The time was 00:00:12.5395606 for: Training
```

I am starting to understand something. These blowups I've been experiencing today might simply be due to numerical inaccuracy. Now I am multiplying by a constant of 0.0001 in the weight update rather than initializing the bias to it, I am getting a completely different result.

Let me do a bit more..."