"5:50pm.
```
The learning rate is 2 ** -10
At iteration 9 the cost and the sign accuracy are 8.913158 and 0.942/1.
At iteration 19 the cost and the sign accuracy are 2.426875 and 0.986/1.
At iteration 29 the cost and the sign accuracy are 1.530533 and 0.99/1.
At iteration 39 the cost and the sign accuracy are 1.325217 and 0.99/1.
At iteration 49 the cost and the sign accuracy are 9.697886 and 0.946/1.
At iteration 59 the cost and the sign accuracy are 2.18108 and 0.984/1.
At iteration 69 the cost and the sign accuracy are 4.345643 and 0.976/1.
At iteration 79 the cost and the sign accuracy are 2.913588 and 0.978/1.
At iteration 89 the cost and the sign accuracy are 4.028605 and 0.98/1.
At iteration 99 the cost and the sign accuracy are 1.45979 and 0.994/1.
At iteration 109 the cost and the sign accuracy are 2.507025 and 0.984/1.
At iteration 119 the cost and the sign accuracy are 1.662976 and 0.99/1.
At iteration 129 the cost and the sign accuracy are 1.508139 and 0.99/1.
At iteration 139 the cost and the sign accuracy are 8.110969 and 0.948/1.
At iteration 149 the cost and the sign accuracy are 8.366407 and 0.944/1.
At iteration 159 the cost and the sign accuracy are 7.875268 and 0.952/1.
At iteration 169 the cost and the sign accuracy are 10.32456 and 0.934/1.
At iteration 179 the cost and the sign accuracy are 10.44203 and 0.938/1.
At iteration 189 the cost and the sign accuracy are 10.67688 and 0.934/1.
At iteration 199 the cost and the sign accuracy are 8.419931 and 0.946/1.
```

It is disturbing how sensitive it is to the constant that multiplies theta.

6:15pm. I give up again.

The plan of the last two days would have been a great one if KFAC had actually pulled through. But it is not enough for this architecture. I've determined two key things:

1) KFAC is absolutely useless on the plastic RNN
2) Gradients are extremely sensitive to the parameter multiplying `theta` and KFAC does nothing to combat that.

All this is extremely strange. You know, before I throw in the towel for good, how about let it run for 2k iterations?"