"8:45am. I did more testing. Hard to believe I spent 40m doing this. At any rate I've come to a conclusion - when the multiplier to `theta` is 0.001, KFAC does in fact do something over standard SGD. Unfortunately that something absolutely wrecks the sign accuracy. With 0.0001 neither SGD nor KFAC can move anywhere.
Let me do some more and then I will post the question and wash my hands of this.

9am. When I jack up the learning rates to -3 and set the covariance modifier to something like -4, that is when I get the KFAC to work better than SGD with the 0.0001 multiplier. Invariance to parameterization? What is that?

No this is not acceptable. Let me go back to `learning2`. I am starting to wonder whether I am doing the right thing there...

9:10am. In `learning2`, let me try a minibatch of 16. I want to see if the error goes up...

```
The learning rate is 2 ** -12
Starting timing for: iteration 0
The time was 00:00:24.7094750 for: iteration 0
Training: 2.04563876397863
The learning rate is 2 ** -12
Starting timing for: iteration 1
The time was 00:00:24.7692360 for: iteration 1
Training: 1.74766940944054
The learning rate is 2 ** -12
Starting timing for: iteration 2
The time was 00:00:24.4807742 for: iteration 2
Training: 1.6791075747721
The learning rate is 2 ** -12
Starting timing for: iteration 3
The time was 00:00:24.6600896 for: iteration 3
Training: 1.64323847498052
The learning rate is 2 ** -12
Starting timing for: iteration 4
The time was 00:00:24.3511830 for: iteration 4
Training: 1.62314864890684
```

No, it seems to be working fine. KFAC is working its magic here. I'll try a minibatch size of 4 after that. This will take 10m to finish."