9:15am. Let me make that question.
[D] How do I get KFAC to work with plastic RNNs?

```
[KFAC](https://arxiv.org/abs/1503.05671) works great for me in RL and in supervised learning with both feedforward and recurrent nets. On the RL game I am testing it on in fact it works so great that I would not dare try doing it without a higher order optimizer. For that reason if I had drop one of the plastic RNNs and KFAC I'd drop the plastic RNNs, but I'd prefer to use both.

That having said, it just works so poorly when it comes to optimizing the plastic RNN on the Binary Pattern test and essentially has none of the invariance to parameterization one would expect from a NG method.

I've managed to figure out a few things about the [plastic RNNs](https://arxiv.org/abs/1804.02464) during my playtime with them. The paper makes a big deal about the plasticity coefficients `alpha`, but even without them on the Binary Pattern it turns out if you let the recurrent connections be plastic and repeatedly feed it the same input plus add it to the output it gets over 99% without training (and any parameters.) This is actually significantly better than one would get by training the net - in fact random initializations tend to make things a lot worse.

I am guessing that this auto-associative memory effect would also work for the Image Recontruction task and possibly the Omniglot. I thought I'd just mention it somewhere since I am not in the business of writing papers and this might be a novel discovery about Hebbian learning. At least it certainly was not obvious to me that repeatedly feeding the same input to the recurrent plastic RNN would lead to it trying to reconstruct the inputs without any special effort - I'd have guessed that the network would be needed to be trained for that first.

Now that is all fine and good, but for a realistic use of plastic RNNs one would want modulation of both the weight updates and the plasticity coefficients. When I left the first as a hyperparameter and I tried optimizing just the later I've found that SGD pretty much crushed KFAC at what turned out to be learning the identity.

When it comes to optimizing only the weight update coefficients using a variant of the Oja's rule the picture is more mixed though still quite bad. Here is how the weight update looks like in Spiral code:

    inm theta = (to float 0.0001) * theta
    H + theta * input * out - abs theta * out * out * H

To prevent an instant blowup, I had to multiply `theta` by a small constant such as 0.0001 otherwise the gradients flowing into it would blow up the network. With this multiplier SGD does not budge from its initial value, and to make KFAC do anything I've had to hike the learning rate to extreme values and mess with the learning rates for the covariance matrices.

A second order method would make this unecessary. In fact, my brilliant plan was to use the semi-second order KFAC to stabilize the plastic RNNs and use it for RL tasks, but as they are right now the plastic RNNs are barely usable and would just be a liability. I was fooled by their great performance on associative memory tasks; SGD is not at all capable of optimizing them.

With the prelude out of the way here are the questions.

* Why does KFAC not work with plastic RNNs?
* Is KFAC not supposed to be model agnostic?
* Is there a way of making it work with plastic RNNs? There are some optimization schemes similar to KFAC like the [Zap](https://arxiv.org/abs/1707.03770) algorithm, but I have no reason to try that one. I could also try centering the gradients and the inputs, but those are just random things that I have no reason to expect would make things work.
* Though unlikely, maybe the square error is making things worse and I should replace it with a probabilistic interpretation of squared error (with multiple classifiers). What sort of cost function would be appropriate for that? I've tried the sigmoid cross entropy, but that did not make any difference.
* Are there any alternative to KFAC I could try? [KFGN](https://arxiv.org/abs/1706.03662) comes to mind, but I would not know how to adapt that for plastic RNNs. [CurveBall](https://arxiv.org/abs/1805.08095) would be another option though that one does not do as well in general compared to block diagonal approximations and would require me to add higher order differentiation to the library I am using. KFAC is great in that it only requires me to understand how to keep track and invert covariance matrices which is a much smaller implementation effort.
```

10:50am. I made a post on the ML sub. It took me 1.5 hours it seems to write it.

https://www.reddit.com/r/MachineLearning/comments/9sbm5o/d_how_do_i_get_kfac_to_work_with_plastic_rnns/

And the run finished. Yeah, even with smaller minibatches KFAC still works well, in fact it works even better. So the reason why the Binary Pattern is working incorrectly is simply because the optimizer is unfit. If I went all the way to online learning I'd guess I'd get an even lower error per iteration over the dataset.

```
The learning rate is 2 ** -12
Starting timing for: iteration 0
The time was 00:01:38.1976838 for: iteration 0
Training: 2.02340957935204
The learning rate is 2 ** -12
Starting timing for: iteration 1
The time was 00:01:37.1682787 for: iteration 1
Training: 1.73433448816002
The learning rate is 2 ** -12
Starting timing for: iteration 2
The time was 00:01:37.5477959 for: iteration 2
Training: 1.66499265903759
The learning rate is 2 ** -12
Starting timing for: iteration 3
The time was 00:01:37.1355506 for: iteration 3
Training: 1.62924869180433
The learning rate is 2 ** -12
Starting timing for: iteration 4
The time was 00:01:37.7568428 for: iteration 4
Training: 1.60736278069749
```

10:55am. That was a vanilla RNN. Vanilla is good, vanilla is love. With that question I mentally transfered the responsibility for plastic RNNs onto the ML sub. This is a temporary measure of course. If I get nothing I'll ditch it sure, but I'll keep an eye out for anything that would break the stalemate and make the plastic RNNs usable.

I really had enough of this. I definitely had enough.

I am going to make vanilla RNNs + KFAC my workhorse. I will implement the mutable optimization and that will perfect them. At that point there will be no more places left to improve upon them. I was really a long journey to get thus far. I could have done all of this a lot earlier if I had the benefit of hindsight, but I can only play the hands I've been given.

11am. I think I will take a break here.

I think the things that are going to come after this will be my long awaited rest. I will free myself of worrying about things I cannot control like those plastic RNNs and optimizers for them and slowly take up the blade and start the crusade to draw out the power of algorithms in the real world.

KFAC + vanilla RNNs will have to suffice for the time being. I'll go forward with that assumption. I've educated myself enough, that is one thing I am certain of.

I want to reward myself for sticking with all of this thus far. By the end of tomorrow I'll know which way to go.

After I get back, I'll implement the mutable optimization for the vanilla RNN and then get the RL stuff operational again. At that point I will aim to implement self play in a RL agent. This is something I can do and have a clear path towards doing. That kind of engineering is my true forte - it certainly is not studying math for natural gradients and researching novel optimizers.