I am still thinking about what I want to do next.
Let me track that paper about denoising RNN states. I think I will put off programming for the rest of the day.

I am starting to form a view on what is going on with modulation. It is not like I need to go for it. There are always alternatives to this.