I had a break to think about it. Let me outline my plan.
First of all, I think I should give modulation a rest and go with my range of learning rates idea.

Earlier this morning I've been considering KFAC alternatives, but ultimately I think KFAC is not wrong. Rather than that, the modulation gradients are extremely uninformative. The same goes for statistics. And this is the key to unraveling this mystery. Thanks to these experiments I've gained an insight into what recurrent Hebbian learning is doing - or at least an aspect of it where the input is added to the output.

Rather than target modulation directly which is a sure way to turn a network into roadkill, what I should be doing is having the static parts generate targets and adding them to the output of the plastic RNN. This would be generalization of the architecture that merely adds the input to the output.

What I should do is forget about the Binary Pattern test. I completely mastered it and trying to fit networks on it would just mislead me. What I should be doing is testing nets on the poker game.

Let me reiterate - the reason KFAC is not working is because the gradients and the statistics for the modulating updates have enormous variance. Since KFAC works so well in general this is a huge sign saying that I should avoid going down this rabbit hole.

What I am thinking now are connections - feedback connections. I finally have a reason to consider them since I have a reason to let a network get the same input for multiple time steps. I've seen this done with LSTMs for planning to great effect.

Since things turned out like this, rather than mixing plastic and static connections on the same layer, I am thinking that the ideal plastic RNN should be a plastic layer on top of static one and with feedback connections going from plastic to static. And both the plastic and the static RNN should have recurrent connections.

The weight update coefficients should be hyperparameters. That will stabilize the variance of gradients coming from the plastic RNN.

No matter what, I cannot abandon my beliefs about KFAC and I cannot abandon the opportunity that the plastic RNNs bring. I absolutely want them both. I will have KFAC make plasticity work. My original plan was not wrong, I just went at it from the wrong angle.

What I need to do it is test it in a RL setting.

Agents doing self play in poker is the ideal test bed for this. I won't get any benefit from plasticity on static rules based players.

I'll leave the Binary Pattern test which I won't touch again on a separate branch and remove it from master along with plasticity. I'll bring out plasticity again when I get to the self play part of my RL journey. So it has been spoken and so it will be done.