"4:40pm. As expected, I won't be programming today. I am still thinking about it. Come to think of it, actually it is pretty remarkable that the outputs themselves are not unstable. They are not that different from modularity coefficients. But whether it be the original plastic RNN or the MI one, they were perfectly stable. KFAC did bring a definite improvement on the plastic RNN from the paper.
Actually it is pretty remarkable that the Binary Pattern managed to converge with only the states as inputs.

From start to finish, my problem was that I keep taking the plasticity papers too literally. I need to be much more skeptical of the conclusions I draw from them. Not just the ones from Miconi but those papers about evolved plastic RNNs on toy tasks that I had read.

Hopefully I'll settle my thoughts soon.

What I will begin tomorrow is getting RL into shape. I am not going to do the mutable optimization. 10-20% gain is not worth the disadvantages of having unsafe semantics in the ML library. What I have now is very fast already as it is. Tomorrow I will start by cleaning things up and then I will get the poker game operational.

Once I do that I will extend the two players so that their optimization is done across the whole game rather than just the individual hands.

At that point I will bring in self play though I think the game will be too simplistic at that point. But it will be easy enough to point the players at NL Holdem itself. I'll make a proper NL HU SNG game.

There is one point my thoughts are starting to settle on - in order for mutation to make sense in the context of RL, I need reversibility. I should avoid things like adding to a vector and then not reversing it on the backward pass.

Even if I could get a speed boost by not reversing, that is just putting my head through the noose. If I am going to support mutation without reversibility I will need to make a system that marks overwritten tensors as null. This would require a fundamental change to `CudaTensor`, though it would not be a difficult one to make.

Spiral has great power so I could even make it so it is optional for when I do not want this particular optimization.

It is important to decide things like this. When one first starts thinking about things, initially the question is how far can it be pushed. But once it is time to get serious, it is more important to ask how can it be made safe. It is a two step process. Once that is done, understanding is attained.

5:25pm. Well, I've been thinking about this for a while now. I like these states of mind, at least until I let them lead me off a cliff like it did with the modularity experiment. But it is all good and fine in the end as I got decent experience from it.

I think it is time I force myself to unwind.

Despite the short term setback, I am near the finish line. With the modularity removed from consideration and the tensor views done, I have everything I need. Implementing that plastic RNN would only be a half a day's work so there is no point in worrying about it.

Having KFAC is a huge boon as it will allow me to use RNNs without caring about initialization. With it present, I only really need to worry about connectivity of the net, and that is straightforward enough. I feel like I understand the effects of that well enough. From the 2012 onwards, most of the great gains in supervised learning came from understanding connectivity.

I definitely won't get that wrong.

I have my future in the palm of my hand. Backpropagation has significant power. If one can extend the GVFNs so that subtasks can be learned automatically, it will be a major step towards the new era.

Once more I need to believe in myself and go with the full force of those ideas. Spiral's ML library has everything it needs to let me take advantage of that."