"8:05am. I got up a 10m ago and now I am chilling. Despite what I wrote yesterday I am still thinking - thinking about plasticity. I realized something huge that should have been obvious to me yesterday, but it was not for some reason - Oja's rule is not good at weight normalization.
It makes sense if the number of weights is small, but suppose you have a large number of units, then their norm would be huge.

I now realize that this is the reason why Oja's rule is less reversible than weight norm.

I made a mistake - I need to move from Oja's rule to weight norm.

Also...during the entire night I've been obsessing what I should set the learning rate to. It seems that I do need the modulation coefficients after all. I cannot do it like this.

I've gotten no reply to my ML sub question. Well, that might be as expected. I guess I won't get anything after all. I have to do this on my own.

...No, ultimately weight norm by itself does not matter. Whether it is right or wrong depends on the problem. So hence, I do need the `alpha` coefficients.

8:35am. Assuming the weight norm Hebbian update...

Then I do not really need the weight modulation is I use the two piece RNN. The reason for that is simple.

The update is `norm(H + in * out)`.

The update with modulation would be `norm(H + m * in * out)`.

If `in` is calculated by another NN, then I can spare myself the effort and simply fold it into `in`. Problem solved.

8:55am. I would gain literally nothing by having the modulation be something separate. The only exception would be if it were generated by a separate process somehow. Then modulation would act as multiplicative integration. In the brain the dopamine system acts to modulate, but what it does cannot be explained by it simply doing multiplication.

9:05am. Let me make a commit here.

I was going to resume RL today, but that is not going to be happening after all. Most likely, I am going to be working on layer norm. I am going to have to redo it so it uses `init_seq`. It is confusing the way it is done right now and is below my standards. That will take some time.

But first I need to make the closing argument and reaffirm my will."