"9:10am. The final conclusion is thus.
It is true that I could not get modulation to work in the last few days. So it goes for KFAC.

But if the two piece plastic RNN with weight norm whose inputs and outputs are generated by the static part cannot be optimized smoothly with KFAC then I can forget about plasticity for good and look for something else. This is absolute. All the perspectives I can think of show that the experiment that is to come will be decisive. The hypothesis feels perfect and without blind spots.

It is different from last time where I simply felt that I had to have modulation. I now have a reason to not have it in explicit form. The principle of hierarchy is quite something. These plastic RNNs when stacked and introspectively unrolled in time will be quite something.

There are stories of the brain going into feedback loops that amplifies its own firing - introspective unrolling exactly fits the description of that.

Hopefully weight norm itself should stabilize the gradients a little too. With Oja's rule the plastic weights should be absolutely huge and would result in huge gradients.

9:35am. So the plan is thus.

I need to redo layer norm and then I will begin work on the true plastic RNN. Everything else will be put on hold and I will drive the experiment to its conclusion.

It will be done on the Binary Pattern test. The most important thing is that it is stable.

KFAC does an amazing job with the vanilla RNN on the `Char RNN` task. I need it to do the same on the `Binary Pattern`. In order to make the test fair, I think I will just initialize all the parameters to zero. The additive connections should ensure that the gradients flow through to the static parts.

9:50am. Right now I am just charging. Even though I've decided on the course of action, my brain feels like mixing it for a while longer. I'll be done brainstorming at some point, but now I just feel like doing it.

I want to say that I will be taking it easy, but brainstorming is hardly taking it easy. It is quite an active thing.

I guess I'll aim to start coding later in the day."