Wasting time is the one true luxury of life.
What I will do now is finally begin work on layer norm.

I won't be satisfied with simply bringing it in. I will give it the complete AD treatment inside the `init_seq` kernel.

This is because I will have to fuse layer norm with the Oja update. I want to make this as elegant as possible. Doing this will pay off in the future at some point.

I need to perfect the ML library.

It will be a pain in the ass, but let me see how far I can go with it today.