To start things off, let me add this to `init_seq`.
I had some time to think about how I want to do this. I am going to simply add a Seq module into CudaAD with the relevant operations for init_seq.

By necessity this is going to involve some duplication of work, but it is not like that much stuff in CudaAD. I can deal with it.

...

```
                        load=inl (!zip tns) ->
                            assert (tns.dim = length a) "The tensor being loaded must have the inner dimension of length equal to inner dimension given to the kernel."
                            create_items (inl {i} -> tns i .get)
                        store=inl {from=(!zip from) to=(!zip to)} ->
                            assert (to.dim = length a) "The tensor being stored must have the inner dimension of length equal to inner dimension given to the kernel."
                            inner_loop {body=inl {item i} -> to i .set (from item .get)}
```

Let me make `load` and `store` like this. It should not be a problem.

...

(30m later) I am going through the possibilities in my mind. I am almost there. I nearly have the new design in view.

The issue is not making the Seq functions. What I am going to have to do though for making everything easier is refactor the CudaAD module. Once I have that, most of the battle will be won.

Yeah it is true that most of the AD functions are straghtforward, but I do not want to repeat the linking functions needlessly.

Most of what I need for the Seq module has already been made. I just need to draw it out.

Let me push this commit and then I will think a bit more and then start. I will start by ignoring the need to do layer norm and Seq, and instead focus on refactoring CudaAD. I want to get at least that done today.