I had a good thing going with those unary operations, but I am completely stuck here. I cannot factor out the `link` functions.
It is ironic, but it was like this last time. The thing that is giving me the most trouble seems to be loading tensors in the first place.

Let me skip the link functions for now and I will redesign the rest. I should have gone to a different branch. I will not be able to do this today. Well, it is not like anyone other than me is using Spiral so it should be fine if everything is broken for a day or two.