I do not feel like starting yet since it is 7:50am, but I just wanted to say that I figured out everything about plastic RNNs. I understand the secrets of modulation in their entirety, not to mention eligibility traces. The architectures that were floating through my minds for the past two days were incorrect.
It is just like me to figure out the absolute simplest solution last. I think I was going in this direction back when I was starting work on plastic RNNs, but I missed it because of the stupid DP paper.

I am sure of it now the approach those guys are going in is misguided.

Here is the true form of differentiable plasticity:

```
x' = xS
y=o(x(SP + B))`
```

Here S is the static part and is a square (recurrent) matrix whose purpose is modulation. P is the recurrent plastic matrix and B is the recurrent static matrix.

It is important that S be full rank in order for the ideal plastic RNN to be realized. The main purpose of `S` is that `x'=xS` generates inputs for the update of P and thereby realizes the implicit modulation idea.

What I tried previously which was applying modulation to the outputs in addition to the evaluation confusion I introduced by following the DP paper is the reason for the downfall of my idea.

In terms of form, this idea is actually pretty close to the one from the DP paper, but the difference is that I now have the reasoning to strongly bet on it.

In this sort of frame, the optimization process itself can develop eligibility traces based on recurrent connections. I do not have to put them explicitly like in the backpropamine paper.

Also, the key to making this work will be KFAC as expected. I think that in practice the plastic RNN should be at most vanilla RNN in terms of optimization difficulty, and it will give me the edge I need to make this work.

For the Binary Pattern, if I set the modulation weights to the identity and the input part of B to identity I will get my optimal performance on that particular task.

It is amazing how many secrets the linear part hides. I never expected it would be turn out like this.

If I view KFAC through the PRONG prism where the natural gradient reprojection is done directly, the actual RNN is:

```
y = o(UxR(SP + B))
```

That is quite a lot of work being done on the linear part. Will there be more?

Oh, yes, there is one thing. Input centering for PRONG should not be done. Consider the biases as inputs which are always 1. If input centering is done, the bias weights won't be trainable which would be wrong. This falsifies the idea.

I thought I should mention this - I never managed to get centering to do anything while I was testing it on Mnist, and now I have a reason not to do it beyond that.

And extending that thread, I probably should not do gradient centering either. What I have now is ideal for the time being.

Without doubt, KFAC is one of the key optimization ideas in deep learning.

Phew. I think I've said all that I wanted to in this commit. I've been busy during the night it seems.

There has been a lot of back and forth over the last few days, but this clinches it. I can safely say that this is the perfected plastic RNN. Either it will work, or nothing else will work.

It will work without a shadow of a doubt. I've internalized it.

With that done, I need a break again to settle my goal into place. I need to turn my mind from the plastic RNN to the weight norm task.

I admit I might be wasting time with `init_seq` since I am unlikely to ever have a layer with more than 1024 outputs, but layer norm is important enough that it deserves good treatment. I will honor it by giving it the proper implementation it deserves.

I have no doubt that this needs to be done. Even if it will shave time from my life, layer norm was one of the motivators for creating Spiral. It will be done justly.

This sort of indulgence should be allowed and even encouraged.

After I am done with this, I will complete the plastic RNN and finally go into RL. I will leave the task of implementing reversible learning for when I need it later.

I think I had enough of experimentation for one year.

Once I am done with the plastic RNN, my goal will be to figure out how to draw out its true power.

I will state it again - doing RL was always and will be my goal. I am definitely looking forward to getting back to it.

Poker is not the last thing that I want to do, but the first thing. I will master all the games and at the end of it, the power will be mine.