I've slacked for a while. It is finally time to do this thing. I think I will collect all the AD operations in order to make the various function in CudaAD using them work.
Then I will move to work on the `Seq` operations.

I'll leave the link for last. I'll go in the direction of what I was thinking of yesterday in terms of those. I'll just grit my teeth and reimplement them properly. It won't be hard.

Let me do it. Let's see if I can get the library operation again today. I want to finish the perfected plastic RNN in the next few days.