"1:30pm. Now comes `link_adjoint`.
```
    inl link_adjoint b {from to} =
        Struct.iter2 (inl from to ->
            match to with
            | () -> ()
            | _ -> k.iter (inl {item i=a} -> to item .set (from b a .get))
            ) from to
        {
        out=()
        bck=inl _ ->
            Struct.iter2 (inl from to ->
                match to with
                | () -> ()
                | _ -> k.iter (inl {item i=a} -> from b a .set (to item .get))
                ) from to
        }
```

This should suffice. Although I do not have `iter`, that should be easy to add. I thought I would need to add more, but it does not seem like that will be the case.

```
inl link_adjoint_view b x = link_adjoint b {x with from=Struct.map (inl x -> x.view) self}
```

Now I admit that this one gave me pause.

Will the `.view` work? Well, one thing I am sure is that it should. I think I should put merging the standard apply and this on the list of things to do. But it is not a priority at the moment.

1:45pm. Things are going well. The binary operations are next. Right now I am thinking of how I am going to do `redo` and `uter`, but either way those will be easy. Strictly speaking, though since I can't really backprop through anything other than sum, maybe I will...well, I'll make the template for the two and implement `sum`. That is the most straightforward thing to do here.

2:15pm. I had to take a break. Let me resume.

As I said, let me do the binary operations. Then I will do the reductions.

And after that I will be done as hard as it is to believe. There is some testing waiting for me after that though."